pyproject.toml
[project]
name = "hello-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "litellm>=1.80.11",
  "python-dotenv>=1.2.1",
  "smolagents[toolkit]>=1.23.0",
]

[dependency-groups]
dev = [
  "pytest>=9.0.2",
  "ruff>=0.14.10",
  "ty>=0.0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ulib"]

[tool.pyrefly]
project-includes = [
  "**/*.py",
  "**/*.ipynb",
]

[tool.uv]
python-preference = "managed"
prerelease = "if-necessary-or-explicit"
resolution = "highest"
